OHS Base YAML install file
I am trying manually work around https://github.com/spm-devops/transformers/issues/1327 - until this is fixed.

On a VM i can manually install all the required pieces i.e.
OHS Base
OHS 19c Upgrade
OHS 19c latest patches

However manually it is more difficult to complete the OHS config for Curam -  as this has a number of steps including creating SSL certs etc.

To help this I have created a YAML file to just install the Base version of OHS.  This will mean when i choose this version when deploying a VM - it will in stall the base and configure OHS for use with Curam..  I can then manually install the 19c upgrade and latest patches - and OHS should be automatically configured for me (from the base install)

This changeset has a YAML file which just installs the Base version (commented out the patches information).  There is a change needed to the ohs main yaml file - to only Patch OHS when the ohs_version variable is set - which it should be for all other yamls (outside of the base)

Tested this locally on molecule - and it looks to pass for both the Base install and latest 12c fixpacks..

If this works will be first step in trying to solve the above transformers ticket.