Update OHS code to upgrade to 19c Client (#106)
* Update OHS code to upgrade to 19c Client

Up to now we have been installing the 12c Client for the Oracle HTTP Server.  This is already or about to go out of support.  So we need to switch to the supported version i.e. 19C Client..

Initially a ticket was created for this work - https://github.com/spm-devops/transformers/issues/1327

But the code deliever in this PR should cover what is required.

* Update README.md

Update README

---------

Co-authored-by: Martin Fanning <mfanning@merative.com>