Update 19c patch var file - add in new patch/fix zip
When deploying OHS with DB Client 19c version we saw an issue logging into Curam post deployment.

This issue lead to issues decoding content and ultimately resources (e.g. css files) not being downloaded and impacting the visual and usability of Curam.

From investagtion we found errors in the logs

This lead us to find Oracle documentation on the errors and highlighted an extra Patch that looks to be required.

This patch was applied to a server manually and once OHS was restarted we no longer saw issues with Curam.

The changes in this PR adds this new patch and ensures it is applied when deploying