Comment out adding WebLogic Domain cert
Comment out adding WebLogic Domain cert

In OHS 14.1.2 there is a change in what certs need to be added to have trust between OHS and WLS.  Previously this was a shared cert - not the cert is specific to the WebLogic Domain.

We do not create a WebLogic domain in our ansible code.  As this is done as part of running the configure target when we deploy Curam.