Update DB2 Molecule and Github Workflows to be not so version specific
Update DB2 Molecule and Github Workflows to be not so version specific

The DB2 Molecule tests are called by the version they represent.  So whenever a new version is released the current tests are renamed to this new version e.g. recent move from db21212 to db21213.

This rename has caused issues with Git Actions expecting a check to run on the old version despite the PR referencing the newer version..  This means the check has to be by passed to alow the PR to be merged..

To avoid this - this PR renames the Molecule tests and GitWorkflows to be more generic for the ocveral l DB2 version.. So instead of the tests and git workflow being called db21159 for the 11.5.9.0 version of DB2.. We now call these tests db2115 - with the fixpack version specifically called out in the converge and verify files as before..

Same for db21213 - this becomes db2121.. 12.1.3 fixpack is referenced in the files of the test.. and we only need to change these files rather then the test name when a newer fixpack is released..