Add in Patching of the Oracle JDK
Add in Patching of the Oracle JDK which looks to be now released as a seperate Patch to the Databae Patch

This often causes compliancy issues in Qualys - so adding in this extra tasks - only ran if the specific var file has a jdk_patch_filename specified - to add this patch.

Kept it seperate in case there is not JDK Patch.